+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ Version 4.0
 - Extract file system functions from ``utils.py`` into ``file_system.py``
 - Extract identification/naming functions from ``utils.py`` into ``identification.py``
 - Extract action related functions from ``api/__init__.py`` to ``actions.py``
+- ``helpers.{modify,ensure,reject}`` moved to ``structure.py``
 
 
 Current versions